Use setting info to determine setting type
This changes the setting type conversion to check the setting's `type`
attribute from the randomizer code rather than simply converting
everything that looks like a `bool` or `int` to that type.

This fixes a bug where the plando would generate seeds with the setting
`"shopsanity": 0` instead of `"shopsanity": "0"`, which behaves almost
identically except that the extra wallet is still generated. The fact
that the randomizer accepts this plando in the first place is a bug.